+++ src/main/java/net/minecraft/client/renderer/EntityRenderer.java
@@ -594,7 +594,7 @@ public class EntityRenderer implements IResourceManagerReloadListener {
                 GlStateManager.rotate(f2 - entity.rotationPitch, 1.0F, 0.0F, 0.0F);
             }
         } else {
-            GlStateManager.translate(0.0F, 0.0F, 0.1F); // FIXED: Parallax Fix
+            GlStateManager.translate(0.0F, 0.0F, -0.1F); // TODO: Parallax Fix But Without Stretching
         }
 
         if (!this.mc.gameSettings.debugCamEnable) {